+++ dbx/cli.py
@@ -182,6 +182,8 @@ app.command(
 )(destroy)
 
 
+# click app object here is used in the mkdocs.
+# Don't delete it!
 def get_click_app():
     return typer.main.get_command(app)