+++ tests/unit/api/adjuster/test_policy.py
@@ -76,6 +76,34 @@ def policy_mock(mocker: MockerFixture):
                         {"spark_version": {"type": "fixed", "value": "some-other"}}
                         """,
                     },
+                    {
+                        "policy_id": 20,
+                        "name": "policy-with-one-script",
+                        "definition": """
+                        {
+                          "init_scripts.0.dbfs.destination": {
+                            "type": "fixed",
+                            "value": "dbfs://some/path/script.sh"
+                          }
+                        }
+                        """,
+                    },
+                    {
+                        "policy_id": 10,
+                        "name": "policy-with-multiple-scripts",
+                        "definition": """
+                        {
+                          "init_scripts.0.dbfs.destination": {
+                            "type": "fixed",
+                            "value": "dbfs://some/path/script.sh"
+                          },
+                          "init_scripts.1.dbfs.destination": {
+                            "type": "fixed",
+                            "value": "dbfs://some/path/other-script.sh"
+                          }
+                        }
+                        """,
+                    },
                 ]
             }
         ),
@@ -158,3 +186,111 @@ def test_locations(policy_mock):
     ]:
         assert getattr(element, "spark_conf").get("spark.my.conf") == "my_value"
         assert element.policy_id == "1"
+
+
+@pytest.mark.parametrize(
+    "existing_init_scripts, expected",
+    [
+        (
+            [],
+            [
+                {"dbfs": {"destination": "dbfs1"}},
+                {"dbfs": {"destination": "dbfs2"}},
+                {"s3": {"destination": "s31"}},
+                {"s3": {"destination": "s32"}},
+            ],
+        ),
+        (
+            [
+                {"dbfs": {"destination": "dbfs1"}},
+                {"dbfs": {"destination": "dbfs2"}},
+                {"s3": {"destination": "s31"}},
+                {"s3": {"destination": "s32"}},
+            ],
+            [
+                {"dbfs": {"destination": "dbfs1"}},
+                {"dbfs": {"destination": "dbfs2"}},
+                {"s3": {"destination": "s31"}},
+                {"s3": {"destination": "s32"}},
+            ],
+        ),
+        (
+            [
+                {"dbfs": {"destination": "dbfs2"}},
+                {"dbfs": {"destination": "dbfs3"}},
+                {"s3": {"destination": "s32"}},
+                {"s3": {"destination": "s33"}},
+            ],
+            [
+                {"dbfs": {"destination": "dbfs1"}},
+                {"dbfs": {"destination": "dbfs2"}},
+                {"s3": {"destination": "s31"}},
+                {"s3": {"destination": "s32"}},
+                {"dbfs": {"destination": "dbfs3"}},
+                {"s3": {"destination": "s33"}},
+            ],
+        ),
+    ],
+)
+def test_append_init_scripts(existing_init_scripts, expected):
+    policy_init_scripts = [
+        {"dbfs": {"destination": "dbfs1"}},
+        {"dbfs": {"destination": "dbfs2"}},
+        {"s3": {"destination": "s31"}},
+        {"s3": {"destination": "s32"}},
+    ]
+    assert expected == PolicyAdjuster._append_init_scripts(policy_init_scripts, existing_init_scripts)
+
+
+TESTS_WITH_SCRIPTS = yaml.safe_load(
+    """
+environments:
+  default:
+    workflows:
+      - name: "one-script"
+        some_task: "here"
+        new_cluster:
+          spark_version: "some"
+          policy_id: "cluster-policy://policy-with-one-script"
+          init_scripts:
+            - dbfs:
+                destination: "dbfs:/some/script.sh"
+      - name: "multiple-scripts"
+        some_task: "here"
+        new_cluster:
+          spark_version: "some"
+          policy_id: "cluster-policy://policy-with-multiple-scripts"
+          init_scripts:
+            - dbfs:
+                destination: "dbfs:/some/script.sh"
+      - name: "wrong-format"
+        some_task: "here"
+        new_cluster:
+          spark_version: "some"
+          policy_id: "cluster-policy://policy-with-one-script"
+          init_scripts:
+            - dbfs:
+              destination: "dbfs:/some/script.sh"
+"""
+)
+
+
+@pytest.mark.parametrize(
+    "wf_name, amount_or_behaviour", [("one-script", 2), ("multiple-scripts", 3), ("wrong-format", Exception())]
+)
+def test_with_scripts(wf_name, amount_or_behaviour, policy_mock):
+    wf = DeploymentConfig.from_payload(TESTS_WITH_SCRIPTS).get_environment("default").payload.get_workflow(wf_name)
+    core_pkg = Library(whl="/some/local/file")
+    client_mock = MagicMock()
+    _adj = Adjuster(
+        additional_libraries=AdditionalLibrariesProvider(core_package=core_pkg),
+        file_uploader=MagicMock(),
+        api_client=client_mock,
+    )
+    if isinstance(amount_or_behaviour, Exception):
+        with pytest.raises(Exception):
+            _adj.traverse([wf])
+    else:
+        _adj.traverse([wf])
+        assert not wf.new_cluster.policy_id.startswith("cluster-policy://")
+        assert len(getattr(wf.new_cluster, "init_scripts")) == amount_or_behaviour